ref: rekognition用のメソッドを他のスクリプトに移動
