--- src/s3_lambda/const/bot_const.py
@@ -1,4 +1,0 @@
-
-# 戦闘力の倍率
-level = 5500
-target_score = 5000 * 100.0
